# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "급식 배틀 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "서비스 상태",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/schools": {
      "get": {
        "operationId": "searchSchools",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "학교 검색 결과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolSearchResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          },
          "504": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    },
    "/meals": {
      "get": {
        "operationId": "getMeals",
        "parameters": [
          {
            "name": "office_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "school_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "날짜별 중식 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealSearchResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          },
          "504": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "ValidationError": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UpstreamError": {
        "description": "NEIS API 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ]
          }
        }
      },
      "School": {
        "type": "object",
        "required": [
          "office_code",
          "office_name",
          "school_code",
          "name",
          "school_type",
          "location"
        ],
        "properties": {
          "office_code": {
            "type": "string"
          },
          "office_name": {
            "type": "string"
          },
          "school_code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "school_type": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SchoolSearchResponse": {
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "has_next"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/School"
            }
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "has_next": {
            "type": "boolean"
          }
        }
      },
      "Meal": {
        "type": "object",
        "required": [
          "date",
          "menu_items"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "menu_items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "calories": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MealSearchResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meal"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      }
    }
  }
}